+++ src/main/java/com/cburch/hex/Test.java
@@ -36,14 +36,14 @@ import javax.swing.JScrollPane;
 public class Test {
   private static class Model implements HexModel {
     private ArrayList<HexModelListener> listeners = new ArrayList<HexModelListener>();
-    private int[] data = new int[924];
+    private long[] data = new long[924];
 
     public void addHexModelListener(HexModelListener l) {
       listeners.add(l);
     }
 
-    public void fill(long start, long len, int value) {
-      int[] oldValues = new int[(int) len];
+    public void fill(long start, long len, long value) {
+      long[] oldValues = new long[(int) len];
       System.arraycopy(data, (int) (start - 11111), oldValues, 0, (int) len);
       Arrays.fill(data, (int) (start - 11111), (int) len, value);
       for (HexModelListener l : listeners) {
@@ -51,7 +51,7 @@ public class Test {
       }
     }
 
-    public int get(long address) {
+    public long get(long address) {
       return data[(int) (address - 11111)];
     }
 
@@ -71,16 +71,16 @@ public class Test {
       listeners.remove(l);
     }
 
-    public void set(long address, int value) {
-      int[] oldValues = new int[] {data[(int) (address - 11111)]};
+    public void set(long address, long value) {
+      long[] oldValues = new long[] {data[(int) (address - 11111)]};
       data[(int) (address - 11111)] = value & 0x1FF;
       for (HexModelListener l : listeners) {
         l.bytesChanged(this, address, 1, oldValues);
       }
     }
 
-    public void set(long start, int[] values) {
-      int[] oldValues = new int[values.length];
+    public void set(long start, long[] values) {
+      long[] oldValues = new long[values.length];
       System.arraycopy(data, (int) (start - 11111), oldValues, 0, values.length);
       System.arraycopy(values, 0, data, (int) (start - 11111), values.length);
       for (HexModelListener l : listeners) {
